// src/types/candidates.ts
export interface CandidateDataMatching {
  sexo: string;
  estado_civil: string;
  pcd: string;
  nivel_academico: string;
  cidade: string;
  nivel_profissional: string;
  ingles: string;
  espanhol: string;
  outros_idiomas: string;
  titulo_profissional: string;
  conhecimentos_tecnicos: string;
  certificacoes: string;
  outras_certificacoes: string;
  area_atuacao: string;
  cv_candidato: string;
}